workers/requirements.txt
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
structlog==23.2.0
yfinance==0.2.28
pandas==2.1.4
numpy>=1.23.2,<1.24.0
scikit-learn==1.3.2
lightgbm==4.1.0
shap==0.44.0
# interpret==0.4.2  # Commented out due to numpy conflicts
vaderSentiment==3.3.2
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
mlflow==2.8.1
prometheus-client==0.19.0
